+++ Sources/Scenes/InviteFriendsList/InviteFriendsListPresenter.swift
@@ -2,6 +2,7 @@ import Foundation
 
 protocol InviteFriendsListPresenterOutput {
     func present(users: [APIClient.User])
+    func present(success: String)
 }
 
 class InviteFriendsListPresenter: InviteFriendsListInteractorOutput {
@@ -13,5 +14,9 @@ class InviteFriendsListPresenter: InviteFriendsListInteractorOutput {
 
     func didFetch(users: [APIClient.User]) {
         output.present(users: users)
+    }
+
+    func presentSuccess(user: APIClient.User) {
+        output.present(success: user.displayName)
     }
 }