+++ Sources/UI/Home/CreateStoryCell.swift
@@ -4,7 +4,6 @@ import UIKit
 class CreateStoryCell: UICollectionViewCell {
     var profileImage: UIImageView = {
         let view = UIImageView()
-        view.layer.cornerRadius = 64 / 2
         view.translatesAutoresizingMaskIntoConstraints = false
         view.layer.masksToBounds = true
         view.clipsToBounds = true
@@ -13,6 +12,16 @@ class CreateStoryCell: UICollectionViewCell {
     }()
 
     private var image: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.backgroundColor = .brandColor
+        view.clipsToBounds = true
+        view.layer.masksToBounds = true
+
+        return view
+    }()
+
+    private var plus: UIView = {
         let view = UIView()
         view.translatesAutoresizingMaskIntoConstraints = false
         view.backgroundColor = .background
@@ -57,8 +66,15 @@ class CreateStoryCell: UICollectionViewCell {
 
         contentView.translatesAutoresizingMaskIntoConstraints = false
 
-        contentView.addSubview(profileImage)
+        image.addSubview(profileImage)
+
         contentView.addSubview(image)
+        contentView.addSubview(plus)
+
+        NSLayoutConstraint.activate([
+            profileImage.heightAnchor.constraint(equalTo: widthAnchor),
+            profileImage.widthAnchor.constraint(equalTo: widthAnchor),
+        ])
 
         NSLayoutConstraint.activate([
             contentView.leftAnchor.constraint(equalTo: leftAnchor),
@@ -66,20 +82,26 @@ class CreateStoryCell: UICollectionViewCell {
         ])
 
         NSLayoutConstraint.activate([
-            profileImage.topAnchor.constraint(equalTo: contentView.topAnchor),
-            profileImage.leftAnchor.constraint(equalTo: contentView.leftAnchor),
-            profileImage.rightAnchor.constraint(equalTo: contentView.rightAnchor),
-            profileImage.heightAnchor.constraint(equalTo: profileImage.widthAnchor),
-            profileImage.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
-            contentView.bottomAnchor.constraint(equalTo: profileImage.bottomAnchor),
+            image.topAnchor.constraint(equalTo: contentView.topAnchor),
+            image.leftAnchor.constraint(equalTo: contentView.leftAnchor),
+            image.rightAnchor.constraint(equalTo: contentView.rightAnchor),
+            image.heightAnchor.constraint(equalTo: widthAnchor),
+            image.widthAnchor.constraint(equalTo: widthAnchor),
+            image.centerXAnchor.constraint(equalTo: contentView.centerXAnchor),
+            contentView.bottomAnchor.constraint(equalTo: image.bottomAnchor),
         ])
 
         NSLayoutConstraint.activate([
-            image.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: 3),
-            image.rightAnchor.constraint(equalTo: contentView.rightAnchor),
-            image.widthAnchor.constraint(equalToConstant: 26),
-            image.heightAnchor.constraint(equalToConstant: 26),
+            plus.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: 3),
+            plus.rightAnchor.constraint(equalTo: contentView.rightAnchor),
+            plus.widthAnchor.constraint(equalToConstant: 26),
+            plus.heightAnchor.constraint(equalToConstant: 26),
         ])
+    }
+
+    override func layoutSubviews() {
+        super.layoutSubviews()
+        image.layer.cornerRadius = frame.size.width / 2
     }
 
     required init?(coder _: NSCoder) {